+++ src/main/java/com/denizenscript/depenizen/bukkit/bridges/WorldGuardBridge.java
@@ -2,7 +2,6 @@ package com.denizenscript.depenizen.bukkit.bridges;
 
 import com.denizenscript.denizen.objects.CuboidTag;
 import com.denizenscript.denizen.objects.LocationTag;
-import com.denizenscript.denizen.objects.NPCTag;
 import com.denizenscript.denizen.objects.PlayerTag;
 import com.denizenscript.denizen.objects.WorldTag;
 import com.denizenscript.denizencore.DenizenCore;
@@ -14,7 +13,7 @@ import com.denizenscript.depenizen.bukkit.commands.worldguard.RegionCommand;
 import com.denizenscript.depenizen.bukkit.objects.worldguard.WorldGuardRegionTag;
 import com.denizenscript.depenizen.bukkit.properties.worldguard.WorldGuardCuboidProperties;
 import com.denizenscript.depenizen.bukkit.properties.worldguard.WorldGuardLocationProperties;
-import com.denizenscript.depenizen.bukkit.properties.worldguard.WorldGuardNPCProperties;
+import com.denizenscript.depenizen.bukkit.properties.worldguard.WorldGuardNPCExtensions;
 import com.denizenscript.depenizen.bukkit.properties.worldguard.WorldGuardPlayerProperties;
 import com.denizenscript.depenizen.bukkit.properties.worldguard.WorldGuardWorldProperties;
 import com.sk89q.worldedit.bukkit.BukkitAdapter;
@@ -34,7 +33,7 @@ public class WorldGuardBridge extends Bridge {
         PropertyParser.registerProperty(WorldGuardPlayerProperties.class, PlayerTag.class);
         PropertyParser.registerProperty(WorldGuardCuboidProperties.class, CuboidTag.class);
         PropertyParser.registerProperty(WorldGuardWorldProperties.class, WorldTag.class);
-        PropertyParser.registerProperty(WorldGuardNPCProperties.class, NPCTag.class);
+        WorldGuardNPCExtensions.register();
         DenizenCore.commandRegistry.registerCommand(RegionCommand.class);
 
         // <--[tag]
